Makefile: don't require xulrunner for build
Since not all distributions (stern look at Ubuntu) provide the pkgconfig file
for libxul, we do not require it for build.
If it is not available, we just hope that whoever uses the Makefile has set the
XUL_CFLAGS, XUL_LDFLAGS or CXXFLAGS environment variable correctly.
If these variables are set or pkg-config returns the right flags for libxul, the
version can be obtained from the "mozilla-config.h" include file using the
preprocessor.